+++ packages/stacked/example/lib/app/app.router.dart
@@ -6,25 +6,27 @@
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:example/app/custom_route_transition.dart' as _i9;
+import 'package:example/datamodels/clashable_one.dart' as _i12;
+import 'package:example/datamodels/clashable_two.dart' as _i13;
+import 'package:example/ui/bottom_nav/bottom_nav_example.dart' as _i3;
+import 'package:example/ui/bottom_nav/favorites/favorites_view.dart' as _i14;
+import 'package:example/ui/bottom_nav/history/history_view.dart' as _i15;
+import 'package:example/ui/bottom_nav/profile/profile_view.dart' as _i16;
+import 'package:example/ui/details/details_view.dart' as _i5;
+import 'package:example/ui/form/example_form_view.dart' as _i6;
+import 'package:example/ui/home/home_view.dart' as _i2;
+import 'package:example/ui/home/home_viewmodel.dart' as _i10;
+import 'package:example/ui/multiple_futures_example/multiple_futures_example_view.dart'
+    as _i18;
+import 'package:example/ui/nonreactive/nonreactive_view.dart' as _i7;
+import 'package:example/ui/stream_view/stream_counter_view.dart' as _i4;
 import 'package:flutter/cupertino.dart' as _i8;
+import 'package:flutter/material.dart' as _i11;
 import 'package:flutter/material.dart';
-import 'package:flutter/src/foundation/key.dart' as _i10;
+import 'package:stacked/src/code_generation/router_annotation/transitions_builders.dart'
+    as _i17;
 import 'package:stacked/stacked.dart' as _i1;
-import 'package:stacked_services/stacked_services.dart' as _i17;
-
-import '../datamodels/clashable_one.dart' as _i12;
-import '../datamodels/clashable_two.dart' as _i11;
-import '../ui/bottom_nav/bottom_nav_example.dart' as _i3;
-import '../ui/bottom_nav/favorites/favorites_view.dart' as _i13;
-import '../ui/bottom_nav/history/history_view.dart' as _i14;
-import '../ui/bottom_nav/profile/profile_view.dart' as _i15;
-import '../ui/details/details_view.dart' as _i5;
-import '../ui/form/example_form_view.dart' as _i6;
-import '../ui/home/home_view.dart' as _i2;
-import '../ui/multiple_futures_example/multiple_futures_example_view.dart'
-    as _i16;
-import '../ui/nonreactive/nonreactive_view.dart' as _i7;
-import '../ui/stream_view/stream_counter_view.dart' as _i4;
+import 'package:stacked_services/stacked_services.dart' as _i19;
 
 class Routes {
   static const homeView = '/';
@@ -65,7 +67,12 @@ class StackedRouter extends _i1.RouterBase {
         orElse: () => HomeViewArguments(),
       );
       return MaterialPageRoute<dynamic>(
-        builder: (context) => _i2.HomeView(key: args.key, title: args.title),
+        builder: (context) => _i2.HomeView(
+            key: args.key,
+            title: args.title,
+            isLoggedIn: args.isLoggedIn,
+            clashableGetter: args.clashableGetter,
+            homeType: args.homeType),
         settings: data,
       );
     },
@@ -116,25 +123,36 @@ class StackedRouter extends _i1.RouterBase {
 }
 
 class HomeViewArguments {
-  const HomeViewArguments({this.key, this.title});
+  const HomeViewArguments(
+      {this.key,
+      this.title = 'hello',
+      this.isLoggedIn = false,
+      this.clashableGetter,
+      this.homeType = _i10.HomeType.appartment});
 
-  final _i10.Key? key;
+  final _i11.Key? key;
 
   final String? title;
+
+  final bool? isLoggedIn;
+
+  final _i12.Clashable Function(String)? clashableGetter;
+
+  final _i10.HomeType homeType;
 }
 
 class StreamCounterViewArguments {
   const StreamCounterViewArguments({this.key, required this.clashableTwo});
 
-  final _i10.Key? key;
+  final _i11.Key? key;
 
-  final List<_i11.Clashable> clashableTwo;
+  final List<_i13.Clashable> clashableTwo;
 }
 
 class DetailsViewArguments {
   const DetailsViewArguments({this.key, required this.name});
 
-  final _i10.Key? key;
+  final _i11.Key? key;
 
   final String name;
 }
@@ -142,7 +160,7 @@ class DetailsViewArguments {
 class ExampleFormViewArguments {
   const ExampleFormViewArguments({this.key, required this.clashableOne});
 
-  final _i10.Key? key;
+  final _i11.Key? key;
 
   final _i12.Clashable clashableOne;
 }
@@ -160,35 +178,32 @@ class BottomNavExampleRoutes {
 class BottomNavExampleRouter extends _i1.RouterBase {
   final _routes = <_i1.RouteDef>[
     _i1.RouteDef(BottomNavExampleRoutes.favoritesView,
-        page: _i13.FavoritesView),
-    _i1.RouteDef(BottomNavExampleRoutes.historyView, page: _i14.HistoryView),
-    _i1.RouteDef(BottomNavExampleRoutes.profileView, page: _i15.ProfileView)
+        page: _i14.FavoritesView),
+    _i1.RouteDef(BottomNavExampleRoutes.historyView, page: _i15.HistoryView),
+    _i1.RouteDef(BottomNavExampleRoutes.profileView, page: _i16.ProfileView)
   ];
 
   final _pagesMap = <Type, _i1.StackedRouteFactory>{
-    _i13.FavoritesView: (data) {
+    _i14.FavoritesView: (data) {
       final args = data.getArgs<NestedFavoritesViewArguments>(
         orElse: () => NestedFavoritesViewArguments(),
       );
       return _i1.buildAdaptivePageRoute<dynamic>(
-        builder: (context) => _i13.FavoritesView(key: args.key, id: args.id),
+        builder: (context) => _i14.FavoritesView(key: args.key, id: args.id),
         settings: data,
       );
     },
-    _i14.HistoryView: (data) {
+    _i15.HistoryView: (data) {
       return PageRouteBuilder<dynamic>(
         pageBuilder: (context, animation, secondaryAnimation) =>
-            const _i14.HistoryView(),
+            const _i15.HistoryView(),
         settings: data,
-        transitionsBuilder: data.transition ??
-            (context, animation, secondaryAnimation, child) {
-              return child;
-            },
+        transitionsBuilder: data.transition ?? _i17.TransitionsBuilders.fadeIn,
       );
     },
-    _i15.ProfileView: (data) {
+    _i16.ProfileView: (data) {
       return _i8.CupertinoPageRoute<dynamic>(
-        builder: (context) => const _i15.ProfileView(),
+        builder: (context) => const _i16.ProfileView(),
         settings: data,
       );
     }
@@ -203,7 +218,7 @@ class BottomNavExampleRouter extends _i1.RouterBase {
 class NestedFavoritesViewArguments {
   const NestedFavoritesViewArguments({this.key, this.id});
 
-  final _i10.Key? key;
+  final _i11.Key? key;
 
   final String? id;
 }
@@ -219,21 +234,21 @@ class FavoritesViewRoutes {
 class FavoritesViewRouter extends _i1.RouterBase {
   final _routes = <_i1.RouteDef>[
     _i1.RouteDef(FavoritesViewRoutes.multipleFuturesExampleView,
-        page: _i16.MultipleFuturesExampleView),
-    _i1.RouteDef(FavoritesViewRoutes.historyView, page: _i14.HistoryView)
+        page: _i18.MultipleFuturesExampleView),
+    _i1.RouteDef(FavoritesViewRoutes.historyView, page: _i15.HistoryView)
   ];
 
   final _pagesMap = <Type, _i1.StackedRouteFactory>{
-    _i16.MultipleFuturesExampleView: (data) {
+    _i18.MultipleFuturesExampleView: (data) {
       return MaterialPageRoute<dynamic>(
-        builder: (context) => const _i16.MultipleFuturesExampleView(),
+        builder: (context) => const _i18.MultipleFuturesExampleView(),
         settings: data,
       );
     },
-    _i14.HistoryView: (data) {
+    _i15.HistoryView: (data) {
       return PageRouteBuilder<dynamic>(
         pageBuilder: (context, animation, secondaryAnimation) =>
-            const _i14.HistoryView(),
+            const _i15.HistoryView(),
         settings: data,
         transitionsBuilder: data.transition ??
             (context, animation, secondaryAnimation, child) {
@@ -249,10 +264,13 @@ class FavoritesViewRouter extends _i1.RouterBase {
   Map<Type, _i1.StackedRouteFactory> get pagesMap => _pagesMap;
 }
 
-extension NavigatorStateExtension on _i17.NavigationService {
+extension NavigatorStateExtension on _i19.NavigationService {
   Future<dynamic> navigateToHomeView(
-      {_i10.Key? key,
-      String? title,
+      {_i11.Key? key,
+      String? title = 'hello',
+      bool? isLoggedIn = false,
+      _i12.Clashable Function(String)? clashableGetter,
+      _i10.HomeType homeType = _i10.HomeType.appartment,
       int? routerId,
       bool preventDuplicates = true,
       Map<String, String>? parameters,
@@ -260,7 +278,12 @@ extension NavigatorStateExtension on _i17.NavigationService {
               BuildContext, Animation<double>, Animation<double>, Widget)?
           transition}) async {
     navigateTo(Routes.homeView,
-        arguments: HomeViewArguments(key: key, title: title),
+        arguments: HomeViewArguments(
+            key: key,
+            title: title,
+            isLoggedIn: isLoggedIn,
+            clashableGetter: clashableGetter,
+            homeType: homeType),
         id: routerId,
         preventDuplicates: preventDuplicates,
         parameters: parameters,
@@ -282,8 +305,8 @@ extension NavigatorStateExtension on _i17.NavigationService {
   }
 
   Future<dynamic> navigateToStreamCounterView(
-      {_i10.Key? key,
-      required List<_i11.Clashable> clashableTwo,
+      {_i11.Key? key,
+      required List<_i13.Clashable> clashableTwo,
       int? routerId,
       bool preventDuplicates = true,
       Map<String, String>? parameters,
@@ -300,7 +323,7 @@ extension NavigatorStateExtension on _i17.NavigationService {
   }
 
   Future<Map<String, List<String>>?> navigateToDetailsView(
-      {_i10.Key? key,
+      {_i11.Key? key,
       required String name,
       int? routerId,
       bool preventDuplicates = true,
@@ -317,7 +340,7 @@ extension NavigatorStateExtension on _i17.NavigationService {
   }
 
   Future<dynamic> navigateToExampleFormView(
-      {_i10.Key? key,
+      {_i11.Key? key,
       required _i12.Clashable clashableOne,
       int? routerId,
       bool preventDuplicates = true,
@@ -349,7 +372,7 @@ extension NavigatorStateExtension on _i17.NavigationService {
   }
 
   Future<dynamic> navigateToNestedFavoritesViewInBottomNavExample(
-      {_i10.Key? key,
+      {_i11.Key? key,
       String? id,
       int? routerId,
       bool preventDuplicates = true,