+++ packages/stacked_generator/lib/src/generators/router/generator/arguments_class/argument_class_builder_helper.dart
@@ -1,7 +1,5 @@
 import 'package:code_builder/code_builder.dart';
 import 'package:stacked_generator/route_config_resolver.dart';
-import 'package:collection/collection.dart';
-import 'package:stacked_generator/src/generators/extensions/string_utils_extension.dart';
 
 class ArgumentClassBuilderHelper {
   final RouteConfig route;
@@ -18,23 +16,9 @@ class ArgumentClassBuilderHelper {
               (b) => b
                 ..modifier = FieldModifier.final$
                 ..name = param.name
-                ..type = param.type.getTypeInsideList == null
-                    ? Reference(
-                        param.type,
-                        param.imports?.firstOrNull,
-                      )
-                    : TypeReference(
-                        (b) => b
-                          ..symbol = param.type.getTypeInsideList?.group(1)
-                          ..types.addAll([
-                            if (param.type.getTypeInsideList != null) ...[
-                              Reference(
-                                param.type.getTypeInsideList?.group(2),
-                                param.imports?.firstOrNull,
-                              ),
-                            ],
-                          ]),
-                      ),
+                ..type = param is FunctionParamConfig
+                    ? param.funRefer
+                    : param.type.refer,
             ))
         .toList();
   }
@@ -56,7 +40,9 @@ class ArgumentClassBuilderHelper {
 
         // Assign default value
         if (param.defaultValueCode != null) {
-          parameterBuilder..defaultTo = literal(param.defaultValueCode!).code;
+          parameterBuilder
+            ..defaultTo =
+                refer(param.defaultValueCode!, param.type.import).code;
         }
 
         // Add required keyword