+++ packages/stacked_generator/lib/src/generators/router/generator/navigate_extension_class/navigate_extension_class_builder_helper.dart
@@ -1,8 +1,6 @@
 import 'package:code_builder/code_builder.dart';
 import 'package:stacked_generator/route_config_resolver.dart';
-import 'package:stacked_generator/src/generators/extensions/string_utils_extension.dart';
 import 'package:stacked_generator/utils.dart';
-import 'package:collection/collection.dart';
 
 class NavigateExtensionClassBuilderHelper {
   Iterable<Method> buildNavigateToExtensionMethods(List<RouteConfig> routes) {
@@ -34,28 +32,14 @@ class NavigateExtensionClassBuilderHelper {
     return Parameter((parameterBuilder) {
       parameterBuilder
         ..name = param.name
-        ..type = param.type.getTypeInsideList == null
-            ? Reference(
-                param.type,
-                param.imports?.firstOrNull,
-              )
-            : TypeReference(
-                (b) => b
-                  ..symbol = param.type.getTypeInsideList?.group(1)
-                  ..types.addAll([
-                    if (param.type.getTypeInsideList != null) ...[
-                      Reference(
-                        param.type.getTypeInsideList?.group(2),
-                        param.imports?.firstOrNull,
-                      ),
-                    ],
-                  ]),
-              )
+        ..type =
+            param is FunctionParamConfig ? param.funRefer : param.type.refer
         ..named = true;
 
       // Assign default value
       if (param.defaultValueCode != null) {
-        parameterBuilder..defaultTo = literal(param.defaultValueCode!).code;
+        parameterBuilder
+          ..defaultTo = refer(param.defaultValueCode!, param.type.import).code;
       }
 
       // Add required keyword