+++ packages/stacked_generator/lib/src/generators/router/generator/router_class/router_class_builder_helper.dart
@@ -68,8 +68,7 @@ class RouterClassBuilderHelper {
     ..symbol = 'List'
     ..types.add(TypeReference((b) => b
       ..symbol = 'RouteDef'
-      ..url = 'package:stacked/stacked.dart'
-      ..isNullable = true)));
+      ..url = 'package:stacked/stacked.dart')));
 
   /// Example
   ///