+++ packages/stacked_generator/lib/src/generators/router/route_config/route_config_factory.dart
@@ -1,7 +1,6 @@
 import 'package:source_gen/source_gen.dart';
 import 'package:stacked_core/stacked_core.dart';
 
-import '../../../../import_resolver.dart';
 import '../models/custom_transition_builder.dart';
 import '../models/route_parameter_config.dart';
 import 'adaptive_route_config.dart';
@@ -34,8 +33,7 @@ class RouteConfigFactory {
     required this.parentClassName,
   });
 
-  RouteConfig fromResolver(
-      ConstantReader stackedRoute, ImportResolver importResolver) {
+  RouteConfig fromResolver(ConstantReader stackedRoute) {
     if (stackedRoute.instanceOf(TypeChecker.fromRuntime(CupertinoRoute))) {
       return CupertinoRouteConfig(
         className: className,