+++ packages/stacked_generator/test/router/navigation_extension_class/navigation_extension_class_test.dart
@@ -1,4 +1,5 @@
 import 'package:code_builder/code_builder.dart';
+import 'package:stacked_generator/resolved_type.dart';
 import 'package:stacked_generator/route_config_resolver.dart';
 import 'package:stacked_generator/src/generators/router/generator/navigate_extension_class/navigate_extension_class_builder.dart';
 import 'package:stacked_generator/src/generators/router/route_config/material_route_config.dart';
@@ -14,17 +15,17 @@ final List<RouteConfig> _routes = [
       className: MapEntry('LoginClass', 'ui/login_class.dart'),
       parameters: [
         RouteParamConfig(
-            isPathParam: false,
-            isQueryParam: false,
-            isPositional: true,
-            name: 'position',
-            type: 'Marker',
-            imports: {'marker.dart'}),
+          isPathParam: false,
+          isQueryParam: false,
+          isPositional: true,
+          name: 'position',
+          type: ResolvedType(name: 'Marker', import: 'marker.dart'),
+        ),
         RouteParamConfig(
           isPathParam: false,
           isQueryParam: false,
           name: 'age',
-          type: 'int',
+          type: ResolvedType(name: 'int'),
         )
       ]),
   MaterialRouteConfig(
@@ -33,25 +34,30 @@ final List<RouteConfig> _routes = [
       className: MapEntry('HomeClass', 'ui/home_class.dart'),
       parameters: [
         RouteParamConfig(
-            isPathParam: false,
-            isQueryParam: false,
-            isPositional: true,
-            name: 'car',
-            type: 'Car',
-            defaultValueCode: 'TOYOTA',
-            imports: {'car.dart'}),
+          isPathParam: false,
+          isQueryParam: false,
+          isPositional: true,
+          name: 'car',
+          type: ResolvedType(name: 'Car', import: 'car.dart'),
+          defaultValueCode: "'TOYOTA'",
+        ),
         RouteParamConfig(
           isPathParam: false,
           isQueryParam: false,
           name: 'age',
-          type: 'int',
+          type: ResolvedType(name: 'int'),
         ),
         RouteParamConfig(
-            isPathParam: false,
-            isQueryParam: false,
-            name: 'markers',
-            imports: {'map.dart'},
-            type: 'List<Marker>'),
+          isPathParam: false,
+          isQueryParam: false,
+          name: 'markers',
+          type: ResolvedType(name: 'List', typeArguments: [
+            ResolvedType(
+              name: 'Marker',
+              import: 'map.dart',
+            )
+          ]),
+        ),
       ])
 ];